Convert modal to page and add breadcrumbs
